Move proposal requests into a helper, Load proposal and results simultaneously
