+++ src/store/modules/app.ts
@@ -1,17 +1,10 @@
-import { getProfiles } from '@/helpers/profile';
 import { getInstance } from '@snapshot-labs/lock/plugins/vue3';
-import { ipfsGet, getScores } from '@snapshot-labs/snapshot.js/src/utils';
-import {
-  getBlockNumber,
-  signMessage
-} from '@snapshot-labs/snapshot.js/src/utils/web3';
+import { getScores } from '@snapshot-labs/snapshot.js/src/utils';
+import { signMessage } from '@snapshot-labs/snapshot.js/src/utils/web3';
 import getProvider from '@snapshot-labs/snapshot.js/src/utils/provider';
-import gateways from '@snapshot-labs/snapshot.js/src/gateways.json';
 import client from '@/helpers/client';
-import { formatProposal, formatProposals, formatSpace } from '@/helpers/utils';
+import { formatProposals, formatSpace } from '@/helpers/utils';
 import { version } from '@/../package.json';
-
-const gateway = process.env.VUE_APP_IPFS_GATEWAY || gateways[0];
 
 const state = {
   init: false,
@@ -157,136 +150,6 @@ const actions = {
       return formatProposals(proposals);
     } catch (e) {
       commit('GET_PROPOSALS_FAILURE', e);
-    }
-  },
-  getProposal: async ({ commit }, { space, id }) => {
-    commit('GET_PROPOSAL_REQUEST');
-    try {
-      console.time('getProposal.data');
-      const response = await Promise.all([ipfsGet(gateway, id)]);
-      console.timeEnd('getProposal.data');
-      let [proposal]: any = response;
-      proposal = formatProposal(proposal);
-      proposal.ipfsHash = id;
-
-      commit('GET_PROPOSAL_SUCCESS');
-      return { proposal };
-    } catch (e) {
-      commit('GET_PROPOSAL_FAILURE', e);
-    }
-  },
-
-  getResults: async ({ commit }, { space, id }) => {
-    commit('GET_RESULTS_REQUEST');
-    try {
-      const provider = getProvider(space.network);
-      console.time('getProposal.results');
-      const response = await Promise.all([
-        ipfsGet(gateway, id),
-        client.request(`${space.key}/proposal/${id}`),
-        getBlockNumber(provider)
-      ]);
-      console.timeEnd('getProposal.results');
-      const [, , blockNumber] = response;
-      let [proposal, votes]: any = response;
-      proposal = formatProposal(proposal);
-      proposal.ipfsHash = id;
-      const voters = Object.keys(votes);
-      const { snapshot } = proposal.msg.payload;
-      const blockTag = snapshot > blockNumber ? 'latest' : parseInt(snapshot);
-
-      /* Get scores */
-      console.time('getProposal.scores');
-      const [scores, profiles]: any = await Promise.all([
-        getScores(
-          space.key,
-          space.strategies,
-          space.network,
-          provider,
-          voters,
-          // @ts-ignore
-          blockTag
-        ),
-        getProfiles([proposal.address, ...voters])
-      ]);
-      console.timeEnd('getProposal.scores');
-      console.log('Scores', scores);
-
-      const authorProfile = profiles[proposal.address];
-      voters.forEach(address => {
-        votes[address].profile = profiles[address];
-      });
-      proposal.profile = authorProfile;
-
-      votes = Object.fromEntries(
-        Object.entries(votes)
-          .map((vote: any) => {
-            vote[1].scores = space.strategies.map(
-              (strategy, i) => scores[i][vote[1].address] || 0
-            );
-            vote[1].balance = vote[1].scores.reduce((a, b: any) => a + b, 0);
-            return vote;
-          })
-          .sort((a, b) => b[1].balance - a[1].balance)
-          .filter(vote => vote[1].balance > 0)
-      );
-
-      /* Get results */
-      const results = {
-        totalVotes: proposal.msg.payload.choices.map(
-          (choice, i) =>
-            Object.values(votes).filter(
-              (vote: any) => vote.msg.payload.choice === i + 1
-            ).length
-        ),
-        totalBalances: proposal.msg.payload.choices.map((choice, i) =>
-          Object.values(votes)
-            .filter((vote: any) => vote.msg.payload.choice === i + 1)
-            .reduce((a, b: any) => a + b.balance, 0)
-        ),
-        totalScores: proposal.msg.payload.choices.map((choice, i) =>
-          space.strategies.map((strategy, sI) =>
-            Object.values(votes)
-              .filter((vote: any) => vote.msg.payload.choice === i + 1)
-              .reduce((a, b: any) => a + b.scores[sI], 0)
-          )
-        ),
-        totalVotesBalances: Object.values(votes).reduce(
-          (a, b: any) => a + b.balance,
-          0
-        )
-      };
-      commit('GET_RESULTS_SUCCESS');
-      return { votes, results };
-    } catch (e) {
-      commit('GET_RESULTS_FAILURE', e);
-    }
-  },
-
-  getPower: async ({ commit }, { space, address, snapshot }) => {
-    commit('GET_POWER_REQUEST');
-    try {
-      const blockNumber = await getBlockNumber(getProvider(space.network));
-      const blockTag = snapshot > blockNumber ? 'latest' : parseInt(snapshot);
-      let scores: any = await getScores(
-        space.key,
-        space.strategies,
-        space.network,
-        getProvider(space.network),
-        [address],
-        // @ts-ignore
-        blockTag
-      );
-      scores = scores.map((score: any) =>
-        Object.values(score).reduce((a, b: any) => a + b, 0)
-      );
-      commit('GET_POWER_SUCCESS');
-      return {
-        scores,
-        totalScore: scores.reduce((a, b: any) => a + b, 0)
-      };
-    } catch (e) {
-      commit('GET_POWER_FAILURE', e);
     }
   }
 };
